+++ src/main/java/chess/domain/game/Ready.java
@@ -3,7 +3,7 @@ package chess.domain.game;
 import chess.domain.board.Board;
 import chess.domain.board.Point;
 import chess.domain.piece.Color;
-import chess.dto.Response;
+import chess.request.Response;
 
 import java.util.List;
