+++ src/test/java/chess/domain/game/RunningTest.java
@@ -5,8 +5,8 @@ import chess.domain.board.Point;
 import chess.domain.piece.Color;
 import chess.domain.piece.King;
 import chess.domain.piece.Queen;
-import chess.dto.BoardAndTurnInfo;
-import chess.dto.Response;
+import chess.request.BoardAndTurnInfo;
+import chess.request.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
